[tests] support for testing the Install, Run targets
Note that the test projects here don't actually consitute a "working" app, it's a white screen because there is no Application or MainPage. It at least tells us that the app successfully deploys and launches.

Later should add some support for querying adb logcat and making sure we don't crash!